+++ tests/unit/test_experiments/test_simulation_with_experiment.py
@@ -90,10 +90,10 @@ class TestSimulationExperiment:
             sol.cycles[0].steps[2]["Voltage [V]"].data, 4.1, rtol=1e-6, atol=1e-5
         )
         np.testing.assert_allclose(
-            sol.cycles[0].steps[3]["Power [W]"].data, 2, rtol=1e-6, atol=1e-5
+            sol.cycles[0].steps[3]["Power [W]"].data, 2, rtol=3e-4, atol=3e-4
         )
         np.testing.assert_allclose(
-            sol.cycles[0].steps[4]["Resistance [Ohm]"].data, 4, rtol=1e-6, atol=1e-5
+            sol.cycles[0].steps[4]["Resistance [Ohm]"].data, 4, rtol=2e-4, atol=6e-4
         )
 
         np.testing.assert_array_equal(
@@ -186,6 +186,16 @@ class TestSimulationExperiment:
         sim = pybamm.Simulation(model, experiment=experiment, parameter_values=param)
         with pytest.raises(pybamm.SolverError, match="skip_ok is True for all steps"):
             sim.solve()
+
+        # Check termination after a skipped step
+        steps = [
+            pybamm.step.Current(2, duration=100.0, skip_ok=False),
+            cc_charge_skip_ok,
+        ]
+        experiment = pybamm.Experiment(steps)
+        sim = pybamm.Simulation(model, experiment=experiment, parameter_values=param)
+        sol = sim.solve()
+        assert sol.termination == "Event exceeded in initial conditions"
 
     def test_all_empty_solution_errors(self):
         model = pybamm.lithium_ion.SPM()
@@ -732,8 +742,10 @@ class TestSimulationExperiment:
         sim = pybamm.Simulation(model, experiment=experiment)
         sim.solve(initial_soc=1)
         np.testing.assert_allclose(
-            sim.solution.cycles[0].last_state.y.full(),
-            sim.solution.cycles[1].steps[-1].first_state.y.full(),
+            sim.solution.cycles[0].last_state.y,
+            sim.solution.cycles[1].steps[-1].first_state.y,
+            atol=1e-15,
+            rtol=1e-15,
         )
 
     def test_solver_error(self):